requirements.txt
# Core AgentCore and Strands dependencies
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.28
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS SDK and utilities
boto3>=1.34.0
botocore>=1.34.0

# LangChain ecosystem (for additional tools and integrations)
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20

# Web search and external APIs
duckduckgo-search>=5.0.0
requests>=2.31.0
httpx>=0.25.0

# OpenTelemetry for observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-langchain>=0.20.0
aws-opentelemetry-distro>=0.10.1

# Data processing and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
numpy>=1.24.0
pandas>=2.0.0